+++ src/services/transformers/users.transformers.ts
@@ -1,0 +1,20 @@
+/*
+ * This file is part of NER's PM Dashboard and licensed under GNU AGPLv3.
+ * See the LICENSE file in the repository root folder for details.
+ */
+
+import { User } from 'utils';
+
+/**
+ * Transforms a user to ensure deep field transformation of date objects.
+ *
+ * @param user Incoming user object supplied by the HTTP response.
+ * @returns Properly transformed user object.
+ */
+export const userTransformer = (user: User) => {
+  return {
+    ...user,
+    firstLogin: new Date(user.firstLogin),
+    lastLogin: new Date(user.lastLogin)
+  };
+};
