+++ src/services/users.api.ts
@@ -6,20 +6,7 @@
 import axios from 'axios';
 import { User } from 'utils';
 import { apiUrls } from '../shared/urls';
-
-/**
- * Transforms a user to ensure deep field transformation of date objects.
- *
- * @param user Incoming user object supplied by the HTTP response.
- * @returns Properly transformed user object.
- */
-export const userTransformer = (user: User) => {
-  return {
-    ...user,
-    firstLogin: new Date(user.firstLogin),
-    lastLogin: new Date(user.lastLogin)
-  };
-};
+import { userTransformer } from './transformers/users.transformers';
 
 /**
  * Fetches all users.
