+++ src/services/work-packages.api.ts
@@ -7,23 +7,7 @@ import axios from 'axios';
 import { WbsNumber, WorkPackage } from 'utils';
 import { wbsPipe } from '../shared/pipes';
 import { apiUrls } from '../shared/urls';
-import { descriptionBulletTransformer } from './projects.api';
-
-/**
- * Transforms a work package to ensure deep field transformation of date objects.
- *
- * @param workPackage Incoming work package object supplied by the HTTP response.
- * @returns Properly transformed work package object.
- */
-export const workPackageTransformer = (workPackage: WorkPackage) => {
-  return {
-    ...workPackage,
-    dateCreated: new Date(workPackage.dateCreated),
-    startDate: new Date(workPackage.startDate),
-    expectedActivities: workPackage.expectedActivities.map(descriptionBulletTransformer),
-    deliverables: workPackage.deliverables.map(descriptionBulletTransformer)
-  };
-};
+import { workPackageTransformer } from './transformers/work-packages.transformers';
 
 /**
  * Fetch all work packages.
